// npm/sdk/src/types/lumen/release/v1/tx.ts
export interface MsgPublishRelease {
  creator: string;
  release: Release | undefined;
}

export interface MsgPublishReleaseResponse {
  id: number;
}

export interface MsgSetEmergency {
  creator: string;
  id: number;
  emergencyOk: boolean;
  /** seconds from now (<= params.max_pending_ttl) */
  emergencyTtl: number;
}

export interface MsgSetEmergencyResponse {
}

export interface MsgValidateRelease {
  authority: string;
  id: number;
}

export interface MsgValidateReleaseResponse {
}

export interface MsgRejectRelease {
  authority: string;
  id: number;
}

export interface MsgRejectReleaseResponse {
}

export interface MsgUpdateParams {
  authority: string;
  params: Params | undefined;
}

export interface MsgUpdateParamsResponse {
}

function createBaseMsgPublishRelease(): MsgPublishRelease {
  return { creator: "", release: undefined };
}

export const MsgPublishRelease: MessageFns<MsgPublishRelease> = {
  encode(message: MsgPublishRelease, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.creator !== "") {
      writer.uint32(10).string(message.creator);
    }
    if (message.release !== undefined) {
      Release.encode(message.release, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgPublishRelease {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgPublishRelease();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.creator = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.release = Release.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgPublishRelease {
    return {
      creator: isSet(object.creator) ? globalThis.String(object.creator) : "",
      release: isSet(object.release) ? Release.fromJSON(object.release) : undefined,
    };
  },

  toJSON(message: MsgPublishRelease): unknown {
    const obj: any = {};
    if (message.creator !== "") {
      obj.creator = message.creator;
    }
    if (message.release !== undefined) {
      obj.release = Release.toJSON(message.release);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgPublishRelease>, I>>(base?: I): MsgPublishRelease {
    return MsgPublishRelease.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgPublishRelease>, I>>(object: I): MsgPublishRelease {
    const message = createBaseMsgPublishRelease();
    message.creator = object.creator ?? "";
    message.release = (object.release !== undefined && object.release !== null)
      ? Release.fromPartial(object.release)
      : undefined;
    return message;
  },
};
// ...
export const MsgSetEmergency: MessageFns<MsgSetEmergency> = {
  encode(message: MsgSetEmergency, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.creator !== "") {
      writer.uint32(10).string(message.creator);
    }
    if (message.id !== 0) {
      writer.uint32(16).uint64(message.id);
    }
    if (message.emergencyOk !== false) {
      writer.uint32(24).bool(message.emergencyOk);
    }
    if (message.emergencyTtl !== 0) {
      writer.uint32(32).int64(message.emergencyTtl);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgSetEmergency {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgSetEmergency();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.creator = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.id = longToNumber(reader.uint64());
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.emergencyOk = reader.bool();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.emergencyTtl = longToNumber(reader.int64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgSetEmergency {
    return {
      creator: isSet(object.creator) ? globalThis.String(object.creator) : "",
      id: isSet(object.id) ? globalThis.Number(object.id) : 0,
      emergencyOk: isSet(object.emergencyOk)
        ? globalThis.Boolean(object.emergencyOk)
        : isSet(object.emergency_ok)
        ? globalThis.Boolean(object.emergency_ok)
        : false,
      emergencyTtl: isSet(object.emergencyTtl)
        ? globalThis.Number(object.emergencyTtl)
        : isSet(object.emergency_ttl)
        ? globalThis.Number(object.emergency_ttl)
        : 0,
    };
  },

  toJSON(message: MsgSetEmergency): unknown {
    const obj: any = {};
    if (message.creator !== "") {
      obj.creator = message.creator;
    }
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    if (message.emergencyOk !== false) {
      obj.emergencyOk = message.emergencyOk;
    }
    if (message.emergencyTtl !== 0) {
      obj.emergencyTtl = Math.round(message.emergencyTtl);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgSetEmergency>, I>>(base?: I): MsgSetEmergency {
    return MsgSetEmergency.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgSetEmergency>, I>>(object: I): MsgSetEmergency {
    const message = createBaseMsgSetEmergency();
    message.creator = object.creator ?? "";
    message.id = object.id ?? 0;
    message.emergencyOk = object.emergencyOk ?? false;
    message.emergencyTtl = object.emergencyTtl ?? 0;
    return message;
  },
};
// ...
export const MsgRejectRelease: MessageFns<MsgRejectRelease> = {
  encode(message: MsgRejectRelease, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.authority !== "") {
      writer.uint32(10).string(message.authority);
    }
    if (message.id !== 0) {
      writer.uint32(16).uint64(message.id);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgRejectRelease {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgRejectRelease();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.authority = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.id = longToNumber(reader.uint64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgRejectRelease {
    return {
      authority: isSet(object.authority) ? globalThis.String(object.authority) : "",
      id: isSet(object.id) ? globalThis.Number(object.id) : 0,
    };
  },

  toJSON(message: MsgRejectRelease): unknown {
    const obj: any = {};
    if (message.authority !== "") {
      obj.authority = message.authority;
    }
    if (message.id !== 0) {
      obj.id = Math.round(message.id);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgRejectRelease>, I>>(base?: I): MsgRejectRelease {
    return MsgRejectRelease.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgRejectRelease>, I>>(object: I): MsgRejectRelease {
    const message = createBaseMsgRejectRelease();
    message.authority = object.authority ?? "";
    message.id = object.id ?? 0;
    return message;
  },
};
// ...
export const MsgUpdateParams: MessageFns<MsgUpdateParams> = {
  encode(message: MsgUpdateParams, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.authority !== "") {
      writer.uint32(10).string(message.authority);
    }
    if (message.params !== undefined) {
      Params.encode(message.params, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): MsgUpdateParams {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseMsgUpdateParams();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.authority = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.params = Params.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): MsgUpdateParams {
    return {
      authority: isSet(object.authority) ? globalThis.String(object.authority) : "",
      params: isSet(object.params) ? Params.fromJSON(object.params) : undefined,
    };
  },

  toJSON(message: MsgUpdateParams): unknown {
    const obj: any = {};
    if (message.authority !== "") {
      obj.authority = message.authority;
    }
    if (message.params !== undefined) {
      obj.params = Params.toJSON(message.params);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<MsgUpdateParams>, I>>(base?: I): MsgUpdateParams {
    return MsgUpdateParams.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<MsgUpdateParams>, I>>(object: I): MsgUpdateParams {
    const message = createBaseMsgUpdateParams();
    message.authority = object.authority ?? "";
    message.params = (object.params !== undefined && object.params !== null)
      ? Params.fromPartial(object.params)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}